requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
chardet==3.0.4
discord.py==1.6.0
google==3.0.0
idna==3.1
idna-ssl==1.1.0
multidict==5.1.0
python-dotenv==0.15.0
soupsieve==2.1
typing-extensions==3.7.4.3
yarl==1.6.3
